+++ RELEASE_NOTES.md
@@ -8,6 +8,8 @@
 
 - The `BatteryPool.power_status` method now streams objects of type `BatteryPoolReport`.  They're identical to the previous `Report` objects, except for the name of the class.
 
+- The `Fuse` class has been moved to the `frequenz.sdk.timeseries` module.
+
 ## New Features
 
 <!-- Here goes the main new features and examples or instructions on how to use them -->